Backend/requirements.txt
asgiref==3.6.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.7
django-admin==2.0.2
django-cors-headers==3.14.0
django-excel-response2==3.0.5
django-filter==22.1
django-six==1.0.5
djangorestframework==3.14.0
et-xmlfile==1.1.0
excel-base==1.0.4
gunicorn==20.1.0
idna==3.4
isoweek==1.3.3
numpy==1.25.0
openpyxl==3.1.2
pandas==2.0.3
passlib==1.7.4
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
stripe==7.5.0
TimeConvert==3.0.13
typing_extensions==4.8.0
tzdata==2022.7
tzlocal==4.3
urllib3==2.1.0
whitenoise==6.4.0
xlwt==1.3.0
